Fix rule copying function
The function expected that the rule directory was created for the
`vagrant` user which happened implicitly with the previous implementation
of `ssh_maxscale`. Now that the `ssh_maxscale` command executes all
commands as root instead of executing only the first one, the test failed
as the owner of the rules directory was root.

Removed the direct command printing and replaced it with functions from
the TestConnections class. Used std::string to make string manipulation a
bit simpler.