Fix replication after mm_mysqlmon
The replication setup should be reset to its original state after the
test. To do this, the checks for slave thread status need to understand
more than the `Yes` and `No` states. All other states are transient which
means that the test should make another attempt to read the status when
one of them is encountered.