Configure kernel parameters for mxs548
The test uses a lot of short connections so it needs to alter the kernel
parameters to better handle it. The test also waits for a short period of
time after the test to allow the connections to die. This will prevent
other tests from failing due to a lack of file descriptors.